+++ dict/headers/hash.h
@@ -1,6 +1,7 @@
 #ifndef DICT_HASH_H
 #define DICT_HASH_H
-#include <stdint.h>
-uint64_t hashed(const char* key);
-uint64_t index_from_hash(const uint64_t hash, const size_t size);
+
+#include "types.h"
+uint64 hashed(const char* key);
+uint64 index_from_hash(const uint64 hash, const uint64 size);
 #endif
