+++ dict/headers/node.h
@@ -1,6 +1,8 @@
 #ifndef DICT_NODE_H
 #define DICT_NODE_H
 #include "errors.h"
+#include "types.h"
+#include <stdlib.h>
 
 typedef void* Element;
 
@@ -11,7 +13,7 @@ typedef struct _Node {
 } *Node;
 
 Node new_node(const char* key, void* value);
-Node node_find(Node* nodes, const char* key, size_t size);
+Node node_find(Node* nodes, const char* key, uint64 size);
 
 Node node_return_self(Node node);
 char* node_return_key(Node node);
